Patched for BBSNET.
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@1628 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/libBBS/utmp.c
@@ -292,6 +292,8 @@ int getnewutmpent(struct user_info *up)
                 && ((now - uentp->freshtime) < 360)) {
                 continue;
             }
+			if (uentp->mode == BBSNET) /* 临时解决办法, flyriver */
+				continue;
             if (uentp->active && uentp->pid && kill(uentp->pid, 0) == -1) {     /*uentp检查 */
                 char buf[STRLEN];
